Add files via upload
This code is of temperature checking by collecting the temperature of num of days from the user and then classifying them into different categories based on the temperature given each day and also calculating the average of total days. It makes the report and saves the file in a .txt file. The code is implemented using class, object, and dictionary.